use Jetbrains annotations instead of findbugs

+++ utils/src/main/java/org/owasp/dependencycheck/utils/FileUtils.java
@@ -20,11 +20,15 @@ package org.owasp.dependencycheck.utils;
 import java.io.*;
 
 import org.apache.commons.io.FilenameUtils;
+import org.apache.commons.lang3.StringUtils;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
 import java.util.UUID;
 import org.apache.commons.lang3.SystemUtils;
+
+import org.jetbrains.annotations.NotNull;
+import org.jetbrains.annotations.Nullable;
 
 /**
  * A collection of utilities for processing information about files.
@@ -60,9 +64,10 @@ public final class FileUtils {
      * @param fileName the file name to retrieve the file extension from.
      * @return the file extension.
      */
-    public static String getFileExtension(String fileName) {
-        final String fileExt = FilenameUtils.getExtension(fileName);
-        return null == fileExt || fileExt.isEmpty() ? null : fileExt.toLowerCase();
+    @Nullable
+    public static String getFileExtension(@NotNull String fileName) {
+        @Nullable final String fileExt = FilenameUtils.getExtension(fileName);
+        return StringUtils.isNoneEmpty(fileExt)? StringUtils.lowerCase(fileExt) : null;
     }
 
     /**
@@ -72,7 +77,13 @@ public final class FileUtils {
      * @param file the File to delete
      * @return true if the file was deleted successfully, otherwise false
      */
-    public static boolean delete(File file) {
+    @NotNull
+    public static boolean delete(@Nullable File file) {
+        if(file == null) {
+            LOGGER.warn("cannot delete null File");
+            return false;
+        }
+
         final boolean success = org.apache.commons.io.FileUtils.deleteQuietly(file);
         if (!success) {
             LOGGER.debug("Failed to delete file: {}; attempting to delete on exit.", file.getPath());
@@ -89,7 +100,8 @@ public final class FileUtils {
      * @throws java.io.IOException thrown when a directory cannot be created within the
      * base directory
      */
-    public static File createTempDirectory(File base) throws IOException {
+    @NotNull
+    public static File createTempDirectory(@Nullable final File base) throws IOException {
         final File tempDir = new File(base, "dctemp" + UUID.randomUUID().toString());
         if (tempDir.exists()) {
             return createTempDirectory(base);
@@ -107,12 +119,9 @@ public final class FileUtils {
      *
      * @return a String containing the bit bucket
      */
+    @NotNull
     public static String getBitBucket() {
-        if (SystemUtils.IS_OS_WINDOWS) {
-            return BIT_BUCKET_WIN;
-        } else {
-            return BIT_BUCKET_UNIX;
-        }
+        return SystemUtils.IS_OS_WINDOWS ? BIT_BUCKET_WIN : BIT_BUCKET_UNIX;
     }
 
     /**
@@ -121,7 +130,7 @@ public final class FileUtils {
      *
      * @param closeable to be closed
      */
-    public static void close(Closeable closeable) {
+    public static void close(@Nullable final Closeable closeable) {
         if (null != closeable) {
             try {
                 closeable.close();
@@ -137,7 +146,8 @@ public final class FileUtils {
      * @param resource path
      * @return the input stream for the given resource
      */
-    public static InputStream getResourceAsStream(final String resource) {
+    @Nullable
+    public static InputStream getResourceAsStream(@NotNull String resource) {
         final ClassLoader classLoader = FileUtils.class.getClassLoader();
         final InputStream inputStream = classLoader != null
                 ? classLoader.getResourceAsStream(resource)
